Get profile command & roles to db save implemented

+++ TaigadevDiscordBot.Core/Extensions/TextChannelExtensions.cs
@@ -1,17 +1,22 @@
 ﻿using System;
 using System.Threading.Tasks;
 
+using Discord;
 using Discord.WebSocket;
 
 namespace TaigadevDiscordBot.Core.Extensions
 {
     public static class TextChannelExtensions
     {
-        public static async Task SendAndRemoveMessageAsync(this ISocketMessageChannel channel, string text, TimeSpan deleteTimespan)
+        public static async Task SendAndRemoveMessageAsync(this ISocketMessageChannel channel, string text, TimeSpan deleteTimespan, Embed embedMessage = null)
         {
-            var message = await channel.SendMessageAsync(text);
-            await Task.Delay(deleteTimespan);
-            await message.DeleteAsync();
+            var message = await channel.SendMessageAsync(text, embed: embedMessage);
+
+            if (deleteTimespan != TimeSpan.MaxValue)
+            {
+                await Task.Delay(deleteTimespan);
+                await message.DeleteAsync();
+            }
         }
     }
 }
